#565 [feature] Change Comment Dto

+++ data/src/main/java/daily/dayo/data/datasource/remote/comment/CommentRequest.kt
@@ -13,7 +13,7 @@ data class CreateCommentRequest(
 
 data class CreateCommentReplyRequest(
     @SerializedName("commentId")
-    val commentId: Int,
+    val commentId: Long,
     @SerializedName("contents")
     val contents: String,
     @SerializedName("postId")
