+++ domain/src/main/java/daily/dayo/domain/usecase/comment/RequestCreatePostCommentReplyUseCase.kt
@@ -7,6 +7,6 @@ import javax.inject.Inject
 class RequestCreatePostCommentReplyUseCase @Inject constructor(
     private val commentRepository: CommentRepository
 ) {
-    suspend operator fun invoke(commentId: Int, contents: String, postId: Int, mentionList: List<MentionUser>) =
+    suspend operator fun invoke(commentId: Long, contents: String, postId: Int, mentionList: List<MentionUser>) =
         commentRepository.requestCreatePostCommentReply(commentId = commentId, contents = contents, postId = postId, mentionList = mentionList)
 }
